Reapply "WIP: Switch from usize to u64"
This reverts commit 3b83843b42d8ee18b36c0e89bcdaddf74575f6e3.

+++ packages/onenote-converter/parser-utils/src/file_api/native_driver.rs
@@ -89,8 +89,8 @@ impl FileApiDriver for FileApiDriverImpl {
 }
 
 impl FileHandle for BufReader<fs::File> {
-    fn byte_length(&self) -> usize {
-        self.get_ref().metadata().map(|m| m.len()).unwrap_or(0) as usize
+    fn byte_length(&self) -> u64 {
+        self.get_ref().metadata().map(|m| m.len()).unwrap_or(0)
     }
 }
